fix: (api & docs) ensure api responses and improve swagger API

+++ app/services/mongoose/platform.js
@@ -1,3 +1,4 @@
+const { isValidObjectId } = require("mongoose");
 const { Platform } = require("../../api/v1/platform/model");
 const {
   BadRequestError,
@@ -5,6 +6,7 @@ const {
   UnauthorizedError,
   NotFoundError,
 } = require("../../errors");
+const { isEmptyOrNull } = require("../../utils/validation");
 
 const createPlatform = async (req) => {
   const { name, url, iconUrl } = req.body;
@@ -26,7 +28,20 @@ const getAllPlatformData = async (req) => {
 
 const getOnePlatformData = async (req) => {
   const { id } = req.params;
-  const result = await Platform.findOne({ _id: id });
+
+  // check is ID valid format
+  const isValidId = isValidObjectId(id);
+
+  if (!isValidId) {
+    throw new BadRequestError("Required platform ID in valid format");
+  }
+
+  const result = await Platform.find({ _id: id });
+
+  // check is result empty
+  if (isEmptyOrNull(result)) {
+    throw new NotFoundError("Platform data not found");
+  }
 
   return result;
 };
@@ -35,9 +50,20 @@ const updatePlatformData = async (req) => {
   const { id } = req.params;
   const { name, url, iconUrl } = req.body;
 
-  if (!id) {
-    throw new BadRequestError("ID is required");
+  // check is ID valid format
+  const isValidId = isValidObjectId(id);
+
+  if (!isValidId) {
+    throw new BadRequestError("Required platform ID in valid format");
   }
+
+  // check is platform exist
+  const checkPlatform = await Platform.find({ _id: id})
+
+  if(isEmptyOrNull(checkPlatform)) {
+    throw new NotFoundError("Platform not found")
+  }
+
 
   const result = await Platform.findOneAndUpdate(
     {
@@ -54,20 +80,31 @@ const updatePlatformData = async (req) => {
     }
   );
 
+
   return result;
 };
 
 const deletePlatformData = async (req) => {
-    const { id } = req.params
+  const { id } = req.params;
 
-    const result = await Platform.findOneAndDelete(
-        {
-            _id: id
-        }
-    )
+  // check is ID valid format
+  const isValidId = isValidObjectId(id);
 
-    return result
-}
+  if (!isValidId) {
+    throw new BadRequestError("Required platform ID in valid format");
+  }
+
+  const result = await Platform.findOneAndDelete({
+    _id: id,
+  });
+
+  // check is result empty (data not found)
+  if(isEmptyOrNull(result)) {
+    throw new NotFoundError("Data not found")
+  }
+
+  return result;
+};
 
 module.exports = {
   createPlatform,
